Migrated some tests to use mocha framework. Added test script in package.json.
