+++ .tasks/data/1da5436c7a31d8298c1282d170bda79d5ea6394901973db14efdea97b3f58331
@@ -1,0 +1,11 @@
+{
+  "status": "done",
+  "pending": "0",
+  "estimation": 1,
+  "title": "Add a test script for mocha in package.json ",
+  "tags": "package",
+  "gitStatus": "A ",
+  "id": "1da5436c7a31d8298c1282d170bda79d5ea6394901973db14efdea97b3f58331",
+  "createdAt": 1476380125150,
+  "updatedAt": 1476380174060
+}
