+++ .tasks/data/2e5e99af1aa399704382d177c3b3ab6eb21f1eb300981cd4cc97abc31f149f9f
@@ -1,0 +1,11 @@
+{
+  "status": "done",
+  "pending": "0",
+  "estimation": 1,
+  "title": "Move the existing tests to mocha tests in a directory",
+  "tags": "project",
+  "gitStatus": "A ",
+  "id": "2e5e99af1aa399704382d177c3b3ab6eb21f1eb300981cd4cc97abc31f149f9f",
+  "createdAt": 1476366608248,
+  "updatedAt": 1476380122500
+}
