+++ lib/server.js
@@ -247,6 +247,113 @@ GitTaskServer.prototype.sortFieldsInDir = function (dir) {
 };
 
 /**
+ * Commit and tag after sorting the task fields in the directory given.
+ * @param dir Directory to commit.
+ */
+GitTaskServer.prototype.commitAfterSort = function(dir) {
+    git.git('add ' + dir.quote());
+    git.git('commit -m "Task fields sorted (for git-task pull)." -- ' + dir.quote());
+    var commitId = git.git('rev-parse --short HEAD');
+    git.git('tag sorted_fields_' + commitId + " " + commitId);
+};
+
+/**
+ * Merge mapped remote task with a local task according to a criteria.
+ * @param remoteMappedTask Remote task mapped to local task.
+ * @param localTask Local task corresponding to the remote task.
+ * @param preferRemote Indicates if the remote fields are preferred against the local fields.
+ * @returns {*} Mapped task.
+ */
+GitTaskServer.prototype.mergeTasks = function(remoteMappedTask, localTask, preferRemote) {
+    var result = {};
+    var taskOrder = preferRemote ? [localTask, remoteMappedTask] : [remoteMappedTask, localTask];
+    taskOrder.forEach(function(t) {
+        var keys = Object.keys(t);
+        keys.forEach(function(k) {
+            result[k] = t[k];
+        });
+    });
+    return result;
+};
+
+const CONFLICT_STRATEGY_LATEST_WINS = 'latest-wins';
+const CONFLICT_STRATEGY_LOCAL_WINS = 'local-wins';
+const CONFLICT_STRATEGY_REMOTE_WINS = 'remote-wins';
+const CONFLICT_STRATEGY_MANUAL = 'manual';
+
+GitTaskServer.prototype.CONFLICT_STRATEGIES = {};
+
+/**
+ * Latest wins conflict strategy.
+ * @param localTask Local task.
+ * @param remoteTask Remote mapped task.
+ * @returns {*} Task.
+ */
+GitTaskServer.prototype.CONFLICT_STRATEGIES[CONFLICT_STRATEGY_LATEST_WINS] = function(localTask, remoteTask) {
+    return localTask.updatedAt > remoteTask.updatedAt ? localTask : remoteTask;
+};
+
+/**
+ * Local task wins conflict strategy.
+ * @param localTask Local task.
+ * @param remoteTask Remote mapped task.
+ * @returns {*} Task.
+ */
+GitTaskServer.prototype.CONFLICT_STRATEGIES[CONFLICT_STRATEGY_LOCAL_WINS] = function(localTask, remoteTask) {
+    return localTask;
+};
+
+/**
+ * Local task wins conflict strategy.
+ * @param localTask Local task.
+ * @param remoteTask Remote mapped task.
+ * @returns {*} Task.
+ */
+GitTaskServer.prototype.CONFLICT_STRATEGIES[CONFLICT_STRATEGY_REMOTE_WINS] = function(localTask, remoteTask) {
+    return remoteTask;
+};
+
+/**
+ * Manual strategy.
+ * @param localTask Local task.
+ * @param remoteTask Remote mapped task.
+ * @returns {*} Task.
+ */
+GitTaskServer.prototype.CONFLICT_STRATEGIES[CONFLICT_STRATEGY_MANUAL] = function(localTask, remoteTask) {
+    return remoteTask;
+};
+
+/**
+ * Merge the local tasks with the remote tasks using a defined strategy.
+ * @param localTasks:Array Local tasks array.
+ * @param remoteTasks:Array Remote mapped tasks array.
+ * @param getIdFn: Function Function to get the ID of the remote task - fn(task): String.
+ * @param strategy:String Strategy to use (use the CONFLICT_STRATEGY_XXX consts).
+ * @returns [] Array of merged tasks.
+ */
+GitTaskServer.prototype.mergeAllTasks = function(localTasks, remoteTasks, getIdFn, strategy) {
+    var idHash = {};
+    var result = [];
+    localTasks.forEach(function(localTask) {
+        var id = getIdFn(localTask);
+        if (typeof id === 'undefined') {
+            result.push(localTask);
+            return;
+        }
+        idHash[id] = localTask;
+    });
+    remoteTasks.forEach(function(remoteTask) {
+        var remoteId = getIdFn(remoteTask);
+        var localTask = idHash[remoteId];
+        var mergedTask = (typeof localTask === 'undefined') ?
+            remoteTask :
+            this.CONFLICT_STRATEGIES[strategy](localTask, remoteTask);
+        result.push(mergedTask);
+    });
+    return result;
+};
+
+/**
  * Read the items from disk.
  * @returns {Array} Item array.
  */
