+++ package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha test/*.js test/**/*.js"
   },
   "repository": {
     "type": "git",
@@ -14,6 +14,10 @@
   },
   "bin": {
       "git-task": "./bin/git-task"
+  },
+  "devDependencies": {
+    "mocha": "3.1.2",
+    "expect.js": "0.3.1"
   },
   "keywords": [
     "git",
